refactor(auth-plugin): extract validators to src/plugins/auth/validators.ts (CP-4)
Error classes (NoActiveOrganizationError, AdminRequiredError, SuperAdminRequiredError)
and role/permission/subscription/feature validators move to validators.ts.
Keeps logUnauthorizedAccess inline (macro-scoped security log).
